==> cellxgene/00_convert_integrated_seurat.Rmd <==
---
title: "CellXgene data"
author: "Matthew Zatzman"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(Seurat)
library(SeuratDisk)
```

Load normalized counts
```{r}
srt <- readRDS("/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/outs_pre/integrated_seurat.rds")
```

Convert to anndata and go to python
```{r}
SaveH5Seurat(srt, filename = "objects/integrated_seurat.h5Seurat")
Convert("objects/integrated_seurat.h5Seurat", dest = "h5ad")
```
