==> cellxgene/01_spectrum_cohort_counts.ipynb <==
import scanpy as sc
import numpy as np
import scipy
import sys
from tqdm import tqdm

This notebook we generate a normalized count matrix from the integrated seurat file masking all other genes as zeroes. Note this requires >300gb memory.

# Load integrated data
adata = sc.read_h5ad("./objects/integrated_seurat.h5ad")
adata

# 3000 genes in X
adata.X

raw_counts = sc.read_h5ad(filename="/work/shah/isabl_data_lake/analyses/84/78/8478/RNASCP/RNASCP/BUILD_MATRIX/fork0/files/matrix.h5")

raw_counts

# subset raw counts for cells in adata
adata.obs.head()

raw_counts = raw_counts[adata.obs.cell_id]
raw_counts

all(raw_counts.obs.index.tolist() == adata.obs['cell_id'])

# Determine which genes are filtered in the normalized matrix
kept_ids = set(adata.var['gene.ids'])
all_ids = set(raw_counts.var['gene_ids'])

# These need to be flagged in 'var' as 'feature_is_filtered' and have counts zeroed
filtered_ids = all_ids.difference(kept_ids)
len(filtered_ids)

# Index on gene ids which are unique
raw_counts.var['gene_name'] = raw_counts.var.index
raw_counts.var.set_index("gene_ids", inplace=True)
raw_counts.var

# Create sparse matrix of zeros equal in size
# LIL format allows better slicing
raw_counts.X = scipy.sparse.lil_array(raw_counts.X.shape)

# Loop over and replace gene values for the cells
for i, gene in enumerate(tqdm(adata.var['gene.ids'])):
    gene_idx = np.where(raw_counts.var.index == gene)[0]    
    raw_counts.X[:,gene_idx] = adata.X[:,i]

# Convert to sparse format and save out (in case memory blows up and need to reload)
raw_counts.X = scipy.sparse.csr_matrix(raw_counts.X)

raw_counts.X

raw_counts.write_h5ad(filename="objects/00_zeroes_plus_normalized_csr.h5ad", compression = "gzip")

# Reload and initialize raw counts matrix...
raw_countmat = sc.read_h5ad(filename="/work/shah/isabl_data_lake/analyses/84/78/8478/RNASCP/RNASCP/BUILD_MATRIX/fork0/files/matrix.h5")

raw_countmat = raw_countmat[adata.obs.cell_id]

raw_countmat.X

# Initialize raw counts matrix in raw
raw_counts.raw = raw_countmat
raw_counts.raw.X

raw_counts.X = scipy.sparse.csr_matrix(raw_counts.X, dtype="float32")

raw_counts.write_h5ad(filename="objects/01_raw_plus_normalized.h5ad", compression = "gzip")
