==> cellxgene/03_celltypes_seurat_to_h5ad.Rmd <==
---
title: "Convert celltype specific objs"
author: "Matthew Zatzman"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(Seurat)
library(SeuratDisk)
library(glue)
library(googlesheets4)
```


Load global object metadata 
```{r}
metadata <- read.csv(file = "./spectrum_scrna_metadata.csv.gz")

# metadata <- metadata %>%
#   mutate(across(where(is.character), factor))
```

Get object specific descriptions
```{r}
study_desc <- read_sheet("https://docs.google.com/spreadsheets/d/15RWlyM8EfB7CgFGc22nIw0yBErkdHz0EFf0DYBMae2M/edit#gid=910670315", sheet = "study_descriptions")
```


Load allele specific information for the ovarian cancer cells
```{r}
ov_bafs <- read.table(file = "/work/shah/vazquezi/data/transfers/spectrum/results/scrna/v27/ascn/outputs/ascn/cell-type/Ovarian.cancer.cell.tsv", header = TRUE, sep = "\t")

# Reformat each chrarm column
ov_bafs_wide <- ov_bafs %>%
  mutate(BAF = signif(BAF, 2)) %>%
  pivot_wider(id_cols = "cell_id", names_from = "chrarm", values_from = "BAF", names_glue = "BAF_{chrarm}")

# Sort columns names nicely
ov_bafs_wide <- ov_bafs_wide[,c("cell_id", gtools::mixedsort(colnames(ov_bafs_wide)[2:ncol(ov_bafs_wide)]))]

```


List of celltype specific objects
```{r}
ct_objs <- list(
  "Ovarian.cancer.cell" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/Ovarian.cancer.super_processed_filtered_annotated.rds",
  "T.super" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/T.super_processed_filtered_annotated.rds",
  "Myeloid.super" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/Myeloid.super_processed_filtered_annotated.rds", 
  "CD8.T" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/CD8.T_processed_filtered.rds",
  "DCs" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/DCs_processed.rds",
  "Macrophages" = "/work/shah/users/uhlitzf/data/SPECTRUM/freeze/v7/Macrophages_processed.rds"
)
```


Loop over and extract, append/merge global metadata.

Need to do as follows: Counts matrix needs raw counts. Data matrix needs normalized counts plus zeroes for all genes which are non-variable. The var column needs a `feature_is_filtered` column to denote all genes which are non-variable.
```{r}
overwrite = TRUE
for (i in seq_along(ct_objs)) {
  ct = names(ct_objs)[i]
  logger::log_info("{i}: {ct}")
  
  final_out = glue("objects/{ct}.h5ad")
  
  if (file.exists(final_out) & !overwrite) {
    logger::log_info("final out exists for {ct}: {final_out}")
  } else {
    logger::log_info("Converting {ct}")
    obj <- readRDS(ct_objs[[i]])
    
    # Metadata columns to merge
    cols_to_merge <- setdiff(colnames(metadata), colnames(obj@meta.data))
  
    new_meta <- data.frame(cbind(obj[[]], metadata[match(colnames(obj), metadata$cell_id), cols_to_merge]))
    
    # Merge arm level BAFs
    if (ct == "Ovarian.cancer.cell") {
      new_meta <- data.frame(cbind(new_meta, ov_bafs_wide[match(new_meta$cell_id, ov_bafs_wide$cell_id), grepl("BAF", colnames(ov_bafs_wide))]))
    }
    
    # Factors don't behave well in conversion
    new_meta <- new_meta %>%
        mutate(across(where(is.factor), as.character))
      
    obj@meta.data <- new_meta
    
    # Set 'is_primary_data' to false
    obj$is_primary_data <- FALSE
    
    # Set variable genes with the 'feature_is_filtered' flag
    obj@assays$RNA@meta.features$feature_is_filtered <- !obj@assays$RNA@meta.features$vst.variable
    
    # Add study title and schema information
    # Unfortunately these don't propagate to the h5ad...
    Misc(obj, slot = "title") <- pull(study_desc[study_desc$object == ct, 'study_description'])
    Misc(obj, slot = "schema_version") <-  "3.0.0"
    
    # Want to get rid of scale data since it will map this to X, and data to raw.X
    # Whereas we want data in X and counts in raw.X
    obj_diet <- DietSeurat(obj, counts = TRUE, data = TRUE, dimreducs = Reductions(obj), graphs = Graphs(obj))
    
    # Convert to h5ads for final processing within python
    SaveH5Seurat(obj_diet, filename = glue("objects/{ct}.h5Seurat"), overwrite = TRUE)
    Convert(glue("objects/{ct}.h5Seurat"), dest = "h5ad", overwrite = TRUE, counts = TRUE)
  }
}
```
